Extract attach bootstrap helpers into separate file
